Free symlink data when unmounting
Calling the yaffs_readdir() function, used for iterating over objects
existing in a given directory, causes a helper function called
yaffs_check_obj_details_loaded() to be executed down the call chain.
When invoked for a symlink object, yaffs_check_obj_details_loaded()
allocates a copy of the target path of the symlink, as a part of lazy
loading logic.  That copy is never freed (not even during unmounting)
unless the symlink is deleted from the file system.

Since Yaffs objects are not allocated one by one, but rather in blocks
of YAFFS_ALLOCATION_NOBJECTS (100), they are also freed in blocks.  Add
a helper function which iterates over the objects present in a block
when it is about to be deallocated and calls kfree() for pointers that
might not be NULL due to yaffs_check_obj_details_loaded() being called
earlier.

Signed-off-by: Michał Kępień <michal@isc.org>